refactor: move bevy code gen functions into its own mod
