+++ crates/bevy_auto_plugin_proc_macros/src/lib.rs
@@ -71,14 +71,14 @@ pub fn module_auto_add_system(_attr: CompilerStream, input: CompilerStream) -> C
 
 /* Flat File */
 
+use bevy_auto_plugin_shared::bevy_app_code_gen::generate_register_type;
 use bevy_auto_plugin_shared::flat_file::inner::expand_flat_file;
 use bevy_auto_plugin_shared::global::__internal::_plugin_entry_block;
 use bevy_auto_plugin_shared::util::TargetRequirePath;
 use bevy_auto_plugin_shared::{
     AddSystemParams, GlobalAutoPluginDeriveParams, GlobalAutoPluginFnAttributeParams,
     GlobalStructOrEnumAttributeParams, StructOrEnumAttributeParams, default_app_ident, flat_file,
-    generate_register_type, get_unique_ident_for_global_struct_or_enum_attribute,
-    ok_or_return_compiler_error,
+    get_unique_ident_for_global_struct_or_enum_attribute, ok_or_return_compiler_error,
 };
 use proc_macro2::Span;
 use quote::{ToTokens, quote};
