+++ crates/bevy_auto_plugin_shared/src/flat_file/inner.rs
@@ -1,14 +1,14 @@
+use crate::bevy_app_code_gen::{
+    generate_add_events, generate_add_systems, generate_auto_names, generate_init_resources,
+    generate_init_states, generate_register_state_types, generate_register_types,
+};
 use crate::flat_file::attribute::FlatFileArgs;
 use crate::flat_file::file_state::{update_file_state, update_state};
 use crate::util::{
     FnParamMutabilityCheckErrMessages, LocalFile, StructOrEnumRef, TargetData, TargetRequirePath,
     is_fn_param_mutable_reference, resolve_local_file, resolve_path_from_item_or_args,
 };
-use crate::{
-    AddSystemParams, StructOrEnumAttributeParams, generate_add_events, generate_add_systems,
-    generate_auto_names, generate_init_resources, generate_init_states,
-    generate_register_state_types, generate_register_types,
-};
+use crate::{AddSystemParams, StructOrEnumAttributeParams};
 use darling::FromMeta;
 use darling::ast::NestedMeta;
 use proc_macro2::{Ident, Span, TokenStream as MacroStream};
