+++ crates/bevy_auto_plugin_shared/src/lib.rs
@@ -1,5 +1,5 @@
 use crate::type_list::TypeList;
-use crate::util::{PathExt, path_to_string, path_to_string_with_spaces};
+use crate::util::{PathExt, path_to_string_with_spaces};
 use darling::{FromDeriveInput, FromField, FromMeta, FromVariant};
 use proc_macro2::{Ident, Span, TokenStream as MacroStream, TokenStream};
 use quote::quote;
@@ -7,6 +7,7 @@ use std::collections::HashSet;
 use std::hash::Hash;
 use syn::{Attribute, Generics, Path, Type, Visibility, parse2};
 
+pub mod bevy_app_code_gen;
 pub mod flat_file;
 pub mod global;
 pub mod module;
@@ -322,162 +323,6 @@ pub struct AutoPluginContext {
     pub init_states: HashSet<String>,
     pub auto_names: HashSet<String>,
     pub add_systems: HashSet<AddSystemSerializedParams>,
-}
-
-pub fn generate_register_type(app_ident: &Ident, item: String) -> syn::Result<MacroStream> {
-    let item = syn::parse_str::<Path>(&item)?;
-    Ok(quote! {
-        #app_ident.register_type::<#item>();
-    })
-}
-
-pub fn generate_register_types(
-    app_ident: &Ident,
-    items: impl Iterator<Item = String>,
-) -> syn::Result<MacroStream> {
-    let register_types = items
-        .map(|item| generate_register_type(app_ident, item))
-        .collect::<syn::Result<Vec<_>>>()?;
-    Ok(quote! {
-        /// generated register_types
-        {
-            #(#register_types)*
-        }
-    })
-}
-
-pub fn generate_add_event(app_ident: &Ident, item: String) -> syn::Result<MacroStream> {
-    let item = syn::parse_str::<Path>(&item)?;
-    Ok(quote! {
-        #app_ident.add_event::<#item>();
-    })
-}
-
-pub fn generate_add_events(
-    app_ident: &Ident,
-    items: impl Iterator<Item = String>,
-) -> syn::Result<MacroStream> {
-    let add_events = items
-        .map(|item| generate_add_event(app_ident, item))
-        .collect::<syn::Result<Vec<_>>>()?;
-    Ok(quote! {
-        /// generated add_events
-        {
-            #(#add_events)*
-        }
-    })
-}
-
-pub fn generate_init_resource(app_ident: &Ident, item: String) -> syn::Result<MacroStream> {
-    let item = syn::parse_str::<Path>(&item)?;
-    Ok(quote! {
-        #app_ident.init_resource::<#item>();
-    })
-}
-
-pub fn generate_init_resources(
-    app_ident: &Ident,
-    items: impl Iterator<Item = String>,
-) -> syn::Result<MacroStream> {
-    let init_resources = items
-        .map(|item| generate_init_resource(app_ident, item))
-        .collect::<syn::Result<Vec<_>>>()?;
-    Ok(quote! {
-        /// generated init_resources
-        {
-            #(#init_resources)*
-        }
-    })
-}
-
-pub fn generate_auto_name(app_ident: &Ident, item: String) -> syn::Result<MacroStream> {
-    let item = syn::parse_str::<Path>(&item)?;
-    let name = path_to_string(&item, true).replace("::", "");
-    Ok(quote! {
-        #app_ident.register_required_components_with::<#item, Name>(|| Name::new(#name));
-    })
-}
-
-pub fn generate_auto_names(
-    app_ident: &Ident,
-    items: impl Iterator<Item = String>,
-) -> syn::Result<MacroStream> {
-    let auto_names = items
-        .map(|item| generate_auto_name(app_ident, item))
-        .collect::<syn::Result<Vec<_>>>()?;
-    Ok(quote! {
-        /// generated auto_names
-        {
-            #(#auto_names)*
-        }
-    })
-}
-
-pub fn generate_register_state_type(app_ident: &Ident, item: String) -> syn::Result<MacroStream> {
-    let item = syn::parse_str::<Path>(&item)?;
-    Ok(quote! {
-        #app_ident.register_type::<State<#item>>();
-        #app_ident.register_type::<NextState<#item>>();
-    })
-}
-
-pub fn generate_register_state_types(
-    app_ident: &Ident,
-    items: impl Iterator<Item = String>,
-) -> syn::Result<MacroStream> {
-    let register_state_types = items
-        .map(|item| generate_register_state_type(app_ident, item))
-        .collect::<syn::Result<Vec<_>>>()?;
-    Ok(quote! {
-        /// generated register_state_types
-        {
-            #(#register_state_types)*
-        }
-    })
-}
-
-pub fn generate_init_state(app_ident: &Ident, item: String) -> syn::Result<MacroStream> {
-    let item = syn::parse_str::<Path>(&item)?;
-    Ok(quote! {
-        #app_ident.init_state::<#item>();
-    })
-}
-
-pub fn generate_init_states(
-    app_ident: &Ident,
-    items: impl Iterator<Item = String>,
-) -> syn::Result<MacroStream> {
-    let init_states = items
-        .map(|item| generate_init_state(app_ident, item))
-        .collect::<syn::Result<Vec<_>>>()?;
-    Ok(quote! {
-        /// generated init_states
-        {
-            #(#init_states)*
-        }
-    })
-}
-
-pub fn generate_add_system(
-    app_ident: &Ident,
-    item: AddSystemSerializedParams,
-) -> syn::Result<MacroStream> {
-    item.to_tokens(app_ident)
-}
-
-pub fn generate_add_systems(
-    app_ident: &Ident,
-    items: impl Iterator<Item = AddSystemSerializedParams>,
-) -> syn::Result<MacroStream> {
-    let output = items
-        .map(|item| generate_add_system(app_ident, item))
-        .collect::<syn::Result<Vec<_>>>()?;
-    Ok(quote! {
-        /// generated add_systems
-        {
-            #(#output)*
-        }
-    })
 }
 
 pub fn default_app_ident() -> Ident {
