+++ crates/bevy_auto_plugin_shared/src/module/inner.rs
@@ -1,13 +1,13 @@
+use crate::bevy_app_code_gen::{
+    generate_add_events, generate_add_systems, generate_auto_names, generate_init_resources,
+    generate_init_states, generate_register_state_types, generate_register_types,
+};
 use crate::module::attribute::ModuleArgs;
 use crate::util::{
     FnRef, ItemWithAttributeMatch, inject_module, items_with_attribute_macro,
     struct_or_enum_items_with_attribute_macro,
 };
-use crate::{
-    AddSystemParams, AddSystemSerializedParams, AutoPluginAttribute, generate_add_events,
-    generate_add_systems, generate_auto_names, generate_init_resources, generate_init_states,
-    generate_register_state_types, generate_register_types,
-};
+use crate::{AddSystemParams, AddSystemSerializedParams, AutoPluginAttribute};
 use darling::FromMeta;
 use darling::ast::NestedMeta;
 use proc_macro2::{Ident, Span, TokenStream as MacroStream};
